requirements.txt
pandas
psutil
fastparquet
pyarrow
pyiceberg
fsspec
minio
s3fs
pendulum
apache-airflow[minio,s3]